ABM de moto realizada

+++ src/app/Shared/Components/card-vehiculo/card-vehiculo.component.ts
@@ -3,6 +3,7 @@ import { Vehiculo } from '../../../Core/Models/Vehicles';
 import { Router } from '@angular/router';
 import { VehiculoPolimorfico } from '../../../Core/Services/Vehicle/VehiculoService/vehiculo.service';
 import { MotoService } from '../../../Core/Services/Vehicle/MotorBike/MotorbikeService/moto.service';
+import { AutoService } from '../../../Core/Services/Vehicle/Car/CarService/auto.service';
 
 @Component({
   selector: 'app-card-vehiculo',
@@ -13,34 +14,47 @@ import { MotoService } from '../../../Core/Services/Vehicle/MotorBike/MotorbikeS
 export class CardVehiculoComponent {
   vehiculo = input.required<VehiculoPolimorfico>()
   router = inject(Router)
-  motoService=inject(MotoService)
-  vehiculoEliminado=output<string>()
+  motoService = inject(MotoService)
+  autoService = inject(AutoService)
+  vehiculoEliminado = output<string>()
 
   modificar() {
-    if(this.vehiculo().tipoVehiculo==="Auto"){
-    this.router.navigate([`vehiculos/modificarAuto/${this.vehiculo().id}`])
-    }else if(this.vehiculo().tipoVehiculo==="Moto"){
+    if (this.vehiculo().tipoVehiculo === "Auto") {
+      this.router.navigate([`vehiculos/modificarAuto/${this.vehiculo().id}`])
+    } else if (this.vehiculo().tipoVehiculo === "Moto") {
       this.router.navigate([`vehiculos/modificarMoto/${this.vehiculo().id}`])
     }
   }
 
   verDetalleVehiculo() {
     if (this.vehiculo().tipoVehiculo === 'Auto') {
-      this.router.navigate([])
+      this.router.navigate([`vehiculos/detalleAuto/${this.vehiculo().id}`])
     }
-    else if(this.vehiculo().tipoVehiculo==="Moto") {
-      this.router.navigate([`vehiculos/detalle/${this.vehiculo().id}`])
+    else if (this.vehiculo().tipoVehiculo === "Moto") {
+      this.router.navigate([`vehiculos/detalleMoto/${this.vehiculo().id}`])
     }
   }
 
-  eliminarVehiculo(){
-    if(this.vehiculo().tipoVehiculo==="Moto"){
-      this.motoService.deleteMoto(this.vehiculo().id).subscribe({
-        next:()=>{
-          alert("Vehiculo eliminado correctamente")
-        },
-        error:(err)=>console.log("Error al eliminar el vehiculo ", err)
-      })
+  eliminarVehiculo() {
+    const resultado = confirm('¿Estás seguro que querés eliminar esta moto?');
+
+    if (!resultado) {
+      console.log("Cancelado")
+    } else {
+      if (this.vehiculo().tipoVehiculo === "Moto") {
+        this.motoService.deleteMoto(this.vehiculo().id).subscribe({
+          next: () => {
+            console.log("Vehiculo eliminado con exito")
+          },
+          error: (err) => console.log("Error al eliminar el vehiculo ", err)
+        })
+      } else if (this.vehiculo().tipoVehiculo === "Auto") {
+        this.autoService.deleteAuto(this.vehiculo().id).subscribe({
+          next: () => {
+          },
+          error: (err) => console.log("Error al eliminar el vehiculo ", err)
+        })
+      }
     }
   }
